requirements-py311.txt
alembic==1.14.0
altgraph==0.17.2
annotated-types==0.7.0
anyio==4.12.1
attrs==24.3.0
beautifulsoup4==4.12.2
bidict==0.23.1
blinker==1.6.2
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.4.0
click==8.1.7
cloudscraper==1.2.71
cryptography==46.0.5
distro==1.9.0
dnspython==2.7.0
emoji==2.10.1
et_xmlfile==2.0.0
eventlet==0.33.3
exceptiongroup==1.2.2
Flask==2.3.3
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
future==0.18.2
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
google-genai==1.47.0
googleapis-common-protos==1.73.0
greenlet==3.1.1
grpcio==1.78.0
grpcio-status==1.71.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.31.2
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.13.0
loguru==0.7.3
lxml==5.1.0
macholib==1.15.2
Mako==1.3.6
MarkupSafe==2.1.3
numpy==2.0.2
oauthlib==3.3.1
openai==2.30.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pdfminer.six==20251107
pdfplumber==0.11.8
Pillow==10.1.0
pip==24.3.1
playwright==1.58.0
pluggy==1.6.0
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.10.2
pydantic_core==2.27.1
pyee==13.0.1
Pygments==2.20.0
pyparsing==3.3.2
pypdfium2==5.6.0
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2025.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
schedule==1.2.1
selenium==4.27.1
setuptools==58.0.4
simple-websocket==1.1.0
six==1.15.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
stripe==5.4.0
tenacity==9.1.2
tomli==2.4.1
tqdm==4.66.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.15.0
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.3.7
wheel==0.37.0
wsproto==1.2.0
zipp==3.21.0
